requirements.txt
blue==0.9.1
isort==5.12.0
Pillow==10.0.1
tinydb==4.8.0
ttkbootstrap==1.10.1
